requirements.txt
# ITMS Developer Setup - Python Dependencies

# Core requirements
requests>=2.31.0
python-dotenv>=1.0.0

# Monday.com integration
monday>=2.0.1

# GitHub integration (requests already included)

# Development tools
black>=23.0.0
pylint>=2.17.0
flake8>=6.0.0

# Optional but recommended
colorama>=0.4.6
rich>=13.0.0